Started making the final scripts

+++ Assignment-3/CS763DeepLearningHW/sahil/Linear.py
@@ -8,8 +8,8 @@ class Linear():
 		super(Linear, self).__init__()
 		self.input_dim = input_dim
 		self.output_dim = output_dim 
-		self.weight = torch.randn(input_dim,output_dim)*sqrt(2/(input_dim))
-		self.bias = torch.randn(1,output_dim)*sqrt(2/(input_dim+output_dim))
+		self.weight = torch.randn(input_dim,output_dim).type(torch.DoubleTensor)*sqrt(2/(input_dim))
+		self.bias = torch.randn(1,output_dim).type(torch.DoubleTensor)*sqrt(2/(input_dim+output_dim))
 		self.isTrainable = True
 		return
 	def forward(self,input):
